+++ R/multiplacesNA.R
@@ -1,24 +1,32 @@
 #' mergeMultiPlacesNA
-#' @description Merges class information for places with multiple entires.
+#' @description Merges information for dublicated places with.
 #' @param df  data.frame.
 #' @param pos_x numeric - column with x-coordinates
 #' @param pos_y numeric - column with y-coordinates
 #' @param col numeric - column with "class"
+#' @param col_ls list, numeric - list with numeric column position which should be merged in
+#' addition to the "class" column. If col_ls==NULL only the class column will be merged. (see details). Default=NULL.
 #' @return returns only unique places.
 #' @note Only useful for classified data!
-#' @details uses pasted x and y coordinates to search for multiple entries and
-#' merges all entires in class for dublicates
+#' @details soon
 #' @author Andreas Schönberg
 #' @export mergeMultiPlacesNA
 #' @aliases mergeMultiPlacesNA
 #' @examples
-#' # load librarys
-#' require(LinguGeo)
-#' require(spatstat)
-#' require(stringr)
+#' # soon
 
-mergeMultiPlacesNA <- function(df,pos_x,pos_y,col,col_ls){
+mergeMultiPlacesNA <- function(df,pos_x,pos_y,col,col_ls=NULL){
 
+  # check col_ls
+  if(any(col_ls==col)){
+    cat("Class column detect in 'col_ls'. Recycle dublicated class column")
+    col_ls <-col_ls[-which(col_ls==col)]
+  }
+  if(is.null(col_ls)==T) {
+    col_ls <- col
+  } else {
+    col_ls <- c(col_ls,col)
+  }
   # paste x and y coorinates to get duplicates by table
   df$places <- paste0(df[,pos_x],".", df[,pos_y])
   occ <- data.frame(table(df$places))
@@ -26,7 +34,7 @@ mergeMultiPlacesNA <- function(df,pos_x,pos_y,col,col_ls){
   # merge occuriance with df
   df2 <- base::merge(df, occ, by.x = "places",by.y= "Var1",all=T)
   # rename column
-  colnames(df2)[which(colnames(df2)=="Freq")] <- "n"
+  colnames(df2)[which(colnames(df2)=="Freq")] <- "n_places"
   # get unique coordinates
   uc <-unique(df2$places)
 
@@ -34,7 +42,7 @@ mergeMultiPlacesNA <- function(df,pos_x,pos_y,col,col_ls){
   for(n in 1:length(uc)){
 
     # check for duplicates
-    if(df2$n[which(df2$places==uc[n])][1]>1) {
+    if(df2$n_places[which(df2$places==uc[n])][1]>1) {
 
       # loop over n columns
       for (nc in 1:length(col_ls)){
@@ -58,11 +66,10 @@ mergeMultiPlacesNA <- function(df,pos_x,pos_y,col,col_ls){
   # delete rows after all dubplicates are pasted
   for(n in 1:length(uc)){
     # delete all dublicates
-    if(df2$n[which(df2$places==uc[n])][1]>1) {
+    if(df2$n_places[which(df2$places==uc[n])][1]>1) {
       df2 <-df2[-which(df2$places==uc[n])[2:length(df2$places[which(df2$places==uc[n])])],]
     }
   }# end loop 'n' over 1:nrows
-  head(df2)
 
   df2$new <- 999
   # clean up class column duplicates due to merging equal classes and handle NA
@@ -73,21 +80,17 @@ mergeMultiPlacesNA <- function(df,pos_x,pos_y,col,col_ls){
 
     # exact one NA <- NA
     if(any(is.na(k_string))==T && length(k_string)==1 ){
-      print("is NA")
       df2$new[k] <- NA
     # multiple 'NA' without any other class  <- NA
     } else if(unique(k_string)=="NA" && length(unique(k_string))==1){
-      print("is NANANA")
         df2$new[k] <- NA
         # 'NA' mixed with other classes <- remove NA
       } else if(any(k_string=='NA')&& length(k_string)>1){
-        print("is mixed")
         # get all non NA values
         k_string2 <-unique(k_string[-which(k_string=='NA')])
         df2$new[k]  <- paste(unique(sort(k_string2)), collapse = ', ')
       } else {
         # all entires !=NA
-        print("nromal case")
         df2$new[k] <-paste(unique(sort(k_string)), collapse = ', ')}# sort to avoid unequal order of elements
       }