+++ R/plot_shannon.R
@@ -1,7 +1,7 @@
 #' plot_shannon
 #' @description Specialized plot function for the Shannon entropy.
-#' @param df  data.frame in the format provided by 'LinguGeo::calc_shannon' (see details).
-#' @param asd Boolean - If FALSE will plot the shannon entropy, if TRUE will plot the Language Item classes and use alpha based on shannon entropy values. Default=F
+#' @param df  data.frame - in the format provided by 'LinguGeo::calc_shannon' (see details).
+#' @param add_data Boolean - If FALSE will plot the shannon entropy, if TRUE will plot the Language Item classes and use alpha based on shannon entropy values. Default=F
 
 #' @return A plot with the shannon entropy or the language item classes in different colors and the shannon entropy as alpha.
 
@@ -10,18 +10,7 @@
 #' @aliases plot_shannon
 #' @examples
 #' # load librarys
-#' require(LinguGeo)
-#' require(ggplot2)
-#' # load data
-#' coh <- read.csv(system.file("extdata","coh_utm.csv",package = "LinguGeo"))
-#' # take a look
-#' head(coh)
-#' # calculate shannon entropy
-#' df_shn <- calc_shannon(coh)
-#' # plot shannon entropy
-#' plot_shannon(df_shn)
-#' # plot add data
-#' plot_shannon(df_shn,add_data=T)
+
 
 plot_shannon <- function(df,add_data=F){
 # check input